fix(core): move git commit to create-nx-workspace
